requirements.txt
asgiref==3.6.0
Django==4.1.5
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
phonenumbers==8.13.3
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
